+++ crates/ptool/src/lua_api.rs
@@ -36,6 +36,10 @@ fn create_ptool_module(
     let unindent_state = Rc::clone(&world);
     let unindent_fn =
         lua.create_function(move |_, input: String| Ok(unindent_state.borrow().unindent(input)))?;
+    let inspect_state = Rc::clone(&world);
+    let inspect_fn = lua.create_function(move |_, (value, options): (Value, Option<Table>)| {
+        inspect_state.borrow().inspect(value, options)
+    })?;
     let args_module = create_ptool_args_module(lua, Rc::clone(&world), script_name, script_args)?;
     let shell_module = create_ptool_shell_module(lua, Rc::clone(&world))?;
     let http_module = create_ptool_http_module(lua, Rc::clone(&world))?;
@@ -52,6 +56,7 @@ fn create_ptool_module(
     module.set("use", use_fn)?;
     module.set("cd", cd_fn)?;
     module.set("unindent", unindent_fn)?;
+    module.set("inspect", inspect_fn)?;
     module.set("args", args_module)?;
     module.set("sh", shell_module)?;
     module.set("http", http_module)?;